Pieces mostly set up
Added piece logic, turn logic, and a game manager. If we keep this setup, we should call GameManager from the GUI pieces and not handle backend board/piece content in the GUI.

Pieces place, but they aren't on the correct spot in the frontend. Working on it.

Edited Board, Cell, and BoardGUI functions so that they're snake_case ^^